Extract Request Cycle Cache from Record and home it in a new interceptors namespace
