+++ lib/lhs/railtie.rb
@@ -15,7 +15,7 @@ module LHS
 
         def prepare_lhs_request_cycle_cache
           return unless LHS.config.request_cycle_cache_enabled
-          LHS::Record::RequestCycleCache::RequestCycleThreadRegistry.request_id = [Time.now.to_f, request.object_id].join('#')
+          LHS::Interceptors::RequestCycleCache::ThreadRegistry.request_id = [Time.now.to_f, request.object_id].join('#')
         end
 
         def reset_option_blocks
